Restructured to use multiple files
A huge change admittedly, but hopefully will be easier to navigate in the future. 
Also added Copy + Paste for tower priorities while viewing a Tower UI
