+++ manifest.json
@@ -1,6 +1,6 @@
 {
     "name": "QoLChanges",
-    "version_number": "1.3.0",
+    "version_number": "1.4.0",
     "website_url": "https://github.com/TheirOwn/QoLChanges",
     "description": "Adds a bunch of quality of life updates.",
     "dependencies": [
